requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.python.org/simple
anyio==2.2.0; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
bandit==1.7.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
cryptography==3.4.7; python_version >= '3.6'
django-admin-sortable2==1.0
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django==3.2.5
djangorestframework==3.12.4
gitdb==4.0.7; python_version >= '3.4'
gitpython==3.1.18; python_version >= '3.6'
h11==0.12.0; python_version >= '3.6'
httpcore==0.12.3; python_version >= '3.6'
httpx==0.16.1; python_version >= '3.6'
idna==3.2
importlib-metadata==4.6.3; python_version < '3.8'
oauthlib==3.1.1; python_version >= '3.6'
pbr==5.6.0; python_version >= '2.6'
pillow==8.3.1
protobuf==3.17.3
psycopg2-binary==2.9.1
pycparser==2.20; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pysocks==1.7.1
pytz==2021.1
pyyaml==5.4.1
requests-oauthlib==1.3.0
requests[socks]==2.26.0
rfc3986[idna2008]==1.5.0
signalstickers-client==3.1.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==4.0.0; python_version >= '3.5'
sniffio==1.2.0; python_version >= '3.5'
sqlparse==0.4.1; python_version >= '3.5'
stevedore==3.3.0; python_version >= '3.6'
tweepy==3.10.0
typing-extensions==3.10.0.0; python_version < '3.8'
urllib3==1.26.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zipp==3.5.0; python_version >= '3.6'
